Delat upp koden mer och gjort ändringar enligt feedback
Delat upp koden i fler filer, ändrat en del klasser från statefull till stateless. Filtret fungerar som det ska igen. Pubspec är uppdaterad.
