+++ lib/TodoItem.dart
@@ -1,0 +1,8 @@
+
+
+class TodoItem {
+  String item;
+  bool isDone;
+
+  TodoItem(this.item, this.isDone);
+}
